move update_plugins to its own function and edit some comments

+++ moodleUpdateCheck.sh
@@ -1,6 +1,11 @@
 #!/bin/bash
 set -e
-image_version="$APP_VERSION"
+
+cur_image_version="$APP_VERSION"
+
+# indicator files
+update_plugins_path="/bitnami/moodledata/UpdatePlugins"
+update_failed_path="/bitnami/moodledata/UpdateFailed"
 
 # checks if image version(new) is greater than current installed version
 version_greater() {
@@ -12,7 +17,7 @@ version_greater() {
 
     greater_version="$(printf "%s\n%s" "$current" "$new" | sort --version-sort --reverse | head -n 1)"
     if [[ "$current" = "$greater_version" ]]; then echo "Current version is higher, unable to downgrade!"; return 0;
-    elif [[ "$new" = "$greater_version" ]]; then echo "Initializing Moodle $image_version ..."; return 1;
+    elif [[ "$new" = "$greater_version" ]]; then echo "Initializing Moodle $cur_image_version ..."; return 1;
     else echo "Unexpected behaviour, exiting version check"; return 0;
     fi
 }
@@ -22,7 +27,7 @@ cleanup() {
     echo "=== Deleting Moodle download folder ==="
     rm -rf /bitnami/moodledata/updated-moodle
     rm -f /bitnami/moodledata/moodle.tgz
-    if ! [ -f /bitnami/moodledata/UpdatePlugins ];
+    if ! [ -f "$update_plugins_path" ];
     then
         echo "=== Disabling maintenance mode and signaling that Update process is finished ==="
         rm -f /bitnami/moodledata/climaintenance.html
@@ -56,35 +61,21 @@ install_kaltura(){
     echo "=== Kaltura Plugin successfully installed ==="
 }
 
-### Start of main ###
-if [ -f /bitnami/moodledata/UpdateFailed ]; then
-    echo "=== UpdateFailed file exists, indicating failed Update! Please resolve the problem manually ==="
-    echo "=== Removing maintenance and starting previous moodle installation ==="
-
-    rm -f /bitnami/moodledata/climaintenance.html
-    start_moodle
-fi
-
-if [ -f /bitnami/moodledata/UpdatePlugins ]; then
-    echo "=== UpdatePlugins File found, starting Plugin installation ==="
+update_plugins() {
     sleep 5
-    rm -f /bitnami/moodledata/UpdatePlugins
+    rm -f "$update_plugins_path"
     if [[ $ENABLE_KALTURA == "True" ]]; then
         echo "=== Kaltura Flag enabled, installing Kaltura Plugin ==="
         install_kaltura
     fi
-    major_regex="\s*([0-9])+\."
-    minor_regex="\.([0-9]*)\."
-    if [[ $image_version =~ $major_regex ]]; then
-            image_major=${BASH_REMATCH[1]}
-    fi
-    if [[ $image_version =~ $minor_regex ]]; then
-            image_minor=${BASH_REMATCH[1]}
-    fi
+
+    # update moodle plugin list
+    php /moosh/moosh.php plugin-list
+
     plugin_version="$image_major.$image_minor"
     nameRegEx="([0-9a-zA-Z_]*)+\#"
     pathRegEx="\#+([0-9a-zA-Z_/]*)"
-    php /moosh/moosh.php plugin-list
+    
     for plugin in $MOODLE_PLUGINS
     do
     # Get plugin name from the list <pluginName>#<pluginPath>
@@ -98,13 +89,35 @@ if [ -f /bitnami/moodledata/UpdatePlugins ]; then
         then
             plugin_name=${BASH_REMATCH[1]}
         fi
-        echo "=== Looking for Plugin: $plugin_name ==="
+        printf '  Looking for "%s" (%s)... ' "$plugin_name" "$plugin_version"
         if [[ -d /bitnami/moodledata/moodle-backup/$plugin_path ]]
         then
-            php /moosh/moosh.php plugin-install "$plugin_name"
-            echo "Plugin $plugin_name for Moodle Version $plugin_version installed"
+            printf "Found!\n"
+            printf "    Starting install..."
+            if php /moosh/moosh.php plugin-install "$plugin_name"; then
+                printf "Done\n"
+            else 
+                printf "Failed!\n"
+            fi
+        else
+            printf "Not found. Skipping\n"
         fi
     done
+}
+
+### Start of main ###
+if [ -f "$update_failed_path" ]; then
+    echo "=== UpdateFailed file exists, indicating failed Update! Please resolve the problem manually ==="
+    echo "=== Removing maintenance and starting previous moodle installation ==="
+
+    rm -f /bitnami/moodledata/climaintenance.html
+    start_moodle
+fi
+
+if [ -f "$update_plugins_path" ]; then
+    echo "=== UpdatePlugins File found, starting Plugin installation ==="
+    update_plugins
+    
     rm -f /bitnami/moodledata/climaintenance.html
     start_moodle
 fi
@@ -138,7 +151,7 @@ pre_update_version=$installed_version;
 echo "The new Moodle Image version is $APP_VERSION";
 
 # Do version check
-if version_greater "$installed_version" "$image_version";
+if version_greater "$installed_version" "$cur_image_version";
 then
     echo "=== Same Version, skipping Update process and starting Moodle ===";
     start_moodle
@@ -146,11 +159,11 @@ fi
 
 # New version, create required Files
 if ! [ -f /bitnami/moodledata/climaintenance.html ]; then
-    echo "=== Enable Maintenance Mode ==="
+    echo "=== Enabling maintenance mode ==="
     echo '<h1>Sorry, maintenance in progress</h1>' > /bitnami/moodledata/climaintenance.html
     sleep 2
     # The backup is only done once in the first run so we don't accidentally overwrite it
-    echo "=== Taking a Backup ==="
+    echo "=== Creating a backup ==="
     if [ -d "/bitnami/moodledata/moodle-backup" ]; then
         rm -r /bitnami/moodledata/moodle-backup
     fi
@@ -162,50 +175,42 @@ fi
 
 # Wait for the Update Helper Job to disable the Probes and force a pod restart
 if ! [ -f /bitnami/moodledata/CliUpdate ]; then
-    echo "=== Create required Files for Update ==="
+    echo "=== Create required files for update ==="
     touch /bitnami/moodledata/CliUpdate
     sleep 600 # Wait for initial Pod termination by Update-Helper-Job
 fi
 
 # Start of the download step
-echo "=== Creating directory for new Version ==="
+echo "=== Creating directory for new version ==="
 if [ -d "/bitnami/moodledata/updated-moodle" ]; then
     rm -rf /bitnami/moodledata/updated-moodle
 fi
 mkdir /bitnami/moodledata/updated-moodle
 
-echo "=== Starting new Moodle Download of Version $image_version ==="
+echo "=== Starting download of new Moodle version $cur_image_version ==="
 
-# Get Version number for download Link and reuse for later Plugin update
+# Get Version number for download Link and reuse for later plugin update
 major_regex="\s*([0-9])+\."
 minor_regex="\.([0-9]*)\."
-if [[ $image_version =~ $major_regex ]]; then
+if [[ $cur_image_version =~ $major_regex ]]; then
         image_major=${BASH_REMATCH[1]}
 fi
-if [[ $image_version =~ $minor_regex ]]; then
+if [[ $cur_image_version =~ $minor_regex ]]; then
         image_minor=${BASH_REMATCH[1]}
 fi
-if [ ${#image_minor} -lt 2 ];
-then two_digit_image_minor=$(printf "%02d" "$image_minor")
+if [ ${#image_minor} -lt 2 ]; then
+    two_digit_image_minor=$(printf "%02d" "$image_minor")
 fi
 stable_version="${image_major}${two_digit_image_minor}"
 
-# Get Version Number for Plugin update
-# if [[ $installed_version =~ $major_regex ]]; then
-#         installed_major=${BASH_REMATCH[1]}
-# fi
-# if [[ $installed_version =~ $minor_regex ]]; then
-#         installed_minor=${BASH_REMATCH[1]}
-# fi
-
 # Test if the download URL is available
-download_url="https://packaging.moodle.org/stable${stable_version}/moodle-${image_version}.tgz"
-# https://download.moodle.org/download.php/direct/stable${stable_version}/moodle-${image_version}.tgz alternative download url
+download_url="https://packaging.moodle.org/stable${stable_version}/moodle-${cur_image_version}.tgz"
+# https://download.moodle.org/download.php/direct/stable${stable_version}/moodle-${cur_image_version}.tgz alternative download url
 echo "Download URL: ${download_url}"
 url_response=$(curl --write-out '%{response_code}' --head --silent --output /dev/null "${download_url}")
 if ! [ "$url_response" -eq 200 ]; then
     echo "=== Critical error, download link is not working, abort update process. Falling back to old version ==="
-    touch /bitnami/moodledata/UpdateFailed
+    touch "$update_failed_path"
     rm /bitnami/moodledata/climaintenance.html
     rm /bitnami/moodledata/CliUpdate && sleep 2
     start_moodle
@@ -225,7 +230,7 @@ cp -rp /bitnami/moodledata/updated-moodle/* /bitnami/moodle/ && echo "=== New mo
 if [[ -n $MOODLE_PLUGINS ]]
 then
     echo "=== Creating UpdatePlugins to trigger Plugin Installation ==="
-    touch /bitnami/moodledata/UpdatePlugins
+    touch "$update_plugins_path"
 else
     echo "=== MOODLE_PLUGINS environment variable missing, skipping plugin copy step ==="
 fi
@@ -245,12 +250,12 @@ else
     # If no moodle Version was found we fall back to previous version
     echo "=== Update failed, no Moodle Version detected. Falling back to old version ==="
     cp -rp /bitnami/moodledata/moodle-backup/* /bitnami/moodle/ && echo "=== Old moodle version restored to folder ==="
-    touch /bitnami/moodledata/UpdateFailed
+    touch "$update_failed_path"
     sleep 5
     start_moodle
 fi
 
-if [ "$post_update_version" == "$image_version" ]; then
+if [ "$post_update_version" == "$cur_image_version" ]; then
     /bin/cp -p /moodleconfig/config.php /bitnami/moodle/config.php
     echo "=== Update to new Version $post_update_version successful. ==="
     echo "=== Starting cleanup ==="
@@ -259,13 +264,13 @@ if [ "$post_update_version" == "$image_version" ]; then
     start_moodle
 elif [ "$post_update_version" == "$pre_update_version" ]; then
     echo "=== Update failed, old Version still installed ==="
-    touch /bitnami/moodledata/UpdateFailed
+    touch "$update_failed_path"
     cleanup
     sleep 10
     start_moodle
 else
     # Normally we should never end up here
     echo "=== Something went wrong, please check the logs(The installed Moodle version does not equal the previous version or the image version) ==="
-    touch /bitnami/moodledata/UpdateFailed
+    touch "$update_failed_path"
     exit 1;
 fi
